move create_tagalong to fiberassign.targets
